A rethinking of how metrics are defined. Separation and encapsulation ftw.
