+++ lib/nerves_hub_link/extensions/health/default_report.ex
@@ -4,8 +4,14 @@ defmodule NervesHubLink.Extensions.Health.DefaultReport do
   new metadata, metrics and such via config.
   """
   @behaviour NervesHubLink.Extensions.Health.Report
+
   alias NervesHubLink.Extensions.Health.Report
-  require Logger
+
+  @default_metric_sets [
+    NervesHubLink.Extensions.Health.MetricSet.CPU,
+    NervesHubLink.Extensions.Health.MetricSet.Disk,
+    NervesHubLink.Extensions.Health.MetricSet.Memory
+  ]
 
   @impl Report
   def timestamp() do
@@ -36,11 +42,7 @@ defmodule NervesHubLink.Extensions.Health.DefaultReport do
   def metrics() do
     [
       metrics_from_config(),
-      cpu_temperature(),
-      cpu_utilization(),
-      load_averages(),
-      memory(),
-      disk()
+      get_metric_sets()
     ]
     |> Enum.reduce(%{}, &Map.merge/2)
   end
@@ -59,8 +61,22 @@ defmodule NervesHubLink.Extensions.Health.DefaultReport do
     |> Enum.reduce(%{}, &Map.merge/2)
   end
 
+  defp get_metric_sets() do
+    Application.get_env(:nerves_hub_link, :health, [])
+    |> Keyword.get(:metric_sets, @default_metric_sets)
+    |> Enum.map(fn metric_set ->
+      if metric_set in [Default, Defaults] do
+        @default_metric_sets
+      else
+        metric_set
+      end
+    end)
+    |> List.flatten()
+    |> Enum.map(& &1.metrics())
+    |> Enum.reduce(%{}, &Map.merge/2)
+  end
+
   defp vof({mod, fun, args}), do: apply(mod, fun, args)
-  defp vof(fun) when is_function(fun), do: fun.()
   defp vof(val), do: val
 
   defp get_health_config(key, default) do
@@ -69,40 +85,35 @@ defmodule NervesHubLink.Extensions.Health.DefaultReport do
   end
 
   defp metadata_from_config() do
-    get_health_config(:metadata, %{})
-    |> format_metrics_or_metadata()
+    metadata = get_health_config(:metadata, %{})
+
+    for {key, val_or_fun} <- metadata, into: %{} do
+      {inspect(key), vof(val_or_fun)}
+    end
   end
 
   defp metrics_from_config() do
-    get_health_config(:metrics, [])
-    |> format_metrics_or_metadata()
+    metrics = get_health_config(:metrics, %{})
+
+    for {key, val_or_fun} <- metrics, into: %{} do
+      {normalize_key(key), vof(val_or_fun)}
+    end
   end
 
-  defp format_metrics_or_metadata(metrics_or_metadata) do
-    metrics_or_metadata
-    |> List.wrap()
-    |> normalize_metrics_from_config()
-    |> evaluate_metrics()
-    |> merge_metrics_output()
+  defp checks_from_config() do
+    checks = get_health_config(:checks, %{})
+
+    for {key, val_or_fun} <- checks, into: %{} do
+      {normalize_key(key), vof(val_or_fun)}
+    end
   end
 
-  defp normalize_metrics_from_config(metrics) do
-    Enum.map(metrics, fn metric ->
-      cond do
-        is_map(metric) -> Map.to_list(metric)
-        metric -> metric
-      end
-    end)
-    |> List.flatten()
-  end
+  defp connectivity_from_config() do
+    connectivity = get_health_config(:connectivity, %{})
 
-  defp evaluate_metrics(metrics) do
-    Enum.map(metrics, fn metric ->
-      case metric do
-        {key, val_or_fun} -> {normalize_key(key), vof(val_or_fun)}
-        metric -> vof(metric)
-      end
-    end)
+    for {key, val_or_fun} <- connectivity, into: %{} do
+      {inspect(key), vof(val_or_fun)}
+    end
   end
 
   defp normalize_key(key) when not is_binary(key) do
@@ -115,152 +126,6 @@ defmodule NervesHubLink.Extensions.Health.DefaultReport do
       key
     else
       inspect(key)
-    end
-  end
-
-  defp merge_metrics_output(metrics) do
-    Enum.reduce(metrics, %{}, fn metric_set, metrics_acc ->
-      cond do
-        is_map(metric_set) ->
-          Map.merge(metrics_acc, metric_set)
-
-        is_tuple(metric_set) ->
-          Map.put(metrics_acc, elem(metric_set, 0), elem(metric_set, 1))
-
-        true ->
-          Logger.warning("Invalid metric set: #{inspect(metric_set)}")
-          metrics_acc
-      end
-    end)
-  end
-
-  defp checks_from_config() do
-    checks = get_health_config(:checks, %{})
-
-    for {key, val_or_fun} <- checks, into: %{} do
-      {inspect(key), vof(val_or_fun)}
-    end
-  end
-
-  defp connectivity_from_config() do
-    connectivity = get_health_config(:connectivity, %{})
-
-    for {key, val_or_fun} <- connectivity, into: %{} do
-      {inspect(key), vof(val_or_fun)}
-    end
-  end
-
-  @default_temperature_source "/sys/class/thermal/thermal_zone0/temp"
-  defp cpu_temperature() do
-    cond do
-      match?({:ok, _}, File.stat(@default_temperature_source)) ->
-        with {:ok, content} <- File.read(@default_temperature_source),
-             {millidegree_c, _} <- Integer.parse(content) do
-          %{cpu_temp: millidegree_c / 1000}
-        else
-          _ ->
-            %{}
-        end
-
-      match?({:ok, _}, File.stat("/usr/bin/vcgencmd")) ->
-        cpu_temperature_rpi()
-
-      true ->
-        %{}
-    end
-  end
-
-  defp cpu_utilization() do
-    case Application.ensure_all_started(:os_mon) do
-      {:ok, _} ->
-        cpu_util()
-
-      {:error, {:already_started, _}} ->
-        cpu_util()
-
-      _ ->
-        %{}
-    end
-  end
-
-  defp cpu_util() do
-    case :cpu_sup.util([]) do
-      {:all, usage, _, _} ->
-        %{cpu_usage_percent: usage}
-
-      _ ->
-        %{}
-    end
-  end
-
-  defp cpu_temperature_rpi() do
-    case System.cmd("/usr/bin/vcgencmd", ["measure_temp"]) do
-      {result, 0} ->
-        %{"temp" => temp} = Regex.named_captures(~r/temp=(?<temp>[\d.]+)/, result)
-        {temp, _} = Integer.parse(temp)
-        %{cpu_temp: temp}
-
-      _ ->
-        %{}
-    end
-  end
-
-  defp load_averages() do
-    with {:ok, data_str} <- File.read("/proc/loadavg"),
-         [min1, min5, min15, _, _] <- String.split(data_str, " "),
-         {min1, _} <- Float.parse(min1),
-         {min5, _} <- Float.parse(min5),
-         {min15, _} <- Float.parse(min15) do
-      %{load_1min: min1, load_5min: min5, load_15min: min15}
-    else
-      _ -> %{}
-    end
-  end
-
-  defp memory() do
-    {free_output, 0} = System.cmd("free", [])
-    [_title_row, memory_row | _] = String.split(free_output, "\n")
-    [_title_column | memory_columns] = String.split(memory_row)
-    [size_kb, used_kb, _, _, _, _] = Enum.map(memory_columns, &String.to_integer/1)
-    size_mb = round(size_kb / 1000)
-    used_mb = round(used_kb / 1000)
-    used_percent = round(used_mb / size_mb * 100)
-
-    %{mem_size_mb: size_mb, mem_used_mb: used_mb, mem_used_percent: used_percent}
-  rescue
-    _ ->
-      %{}
-  end
-
-  defp disk() do
-    case Application.ensure_all_started(:os_mon) do
-      {:ok, _} ->
-        disk_info()
-
-      {:error, {:already_started, _}} ->
-        disk_info()
-
-      _ ->
-        %{}
-    end
-  end
-
-  defp disk_info() do
-    data =
-      Enum.find(:disksup.get_disk_info(), fn {key, _, _, _} ->
-        key == ~c"/root"
-      end)
-
-    case data do
-      nil ->
-        %{}
-
-      {_, total_kb, available_kb, capacity_percentage} ->
-        %{
-          disk_total_kb: total_kb,
-          disk_available_kb: available_kb,
-          disk_used_percentage: capacity_percentage
-        }
     end
   end
