+++ lib/nerves_hub_link/extensions/health/metric_set.ex
@@ -1,0 +1,7 @@
+defmodule NervesHubLink.Extensions.Health.MetricSet do
+  @moduledoc """
+  Behaviour for implementing a custom metric set to be used in a report.
+  """
+
+  @callback metrics() :: %{String.t() => number()}
+end
